fixed tests

+++ src/plugins/intel_gpu/src/graph/impls/ocl/rms.cpp
@@ -38,7 +38,7 @@ struct rms_impl : typed_primitive_impl_ocl<rms> {
 
         params.inputs.push_back(convert_data_tensor(impl_param.get_input_layout(1)));
         params.epsilon = primitive->epsilon;
-        params.ov_input_rank = primitive->input_rank;
+        params.ov_input_rank = static_cast<int32_t>(primitive->input_rank);
         return params;
     }
